Refactor plugin.go for smaller file size
